backend/requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.9.0
django-debug-toolbar==3.2.2
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pyparsing==2.4.7
pytz==2021.1
regex==2021.9.24
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
